+++ lib/public/components/environments/environmentStatusHistoryComponent.js
@@ -11,14 +11,40 @@
  *  or submit itself to any jurisdiction.
  */
 
-import { environmentStatusHistoryLegendComponent } from './environmentStatusHistoryLegendComponent.js';
 import { tooltip } from '../common/popover/tooltip.js';
+import { coloredEnvironmentStatusComponent } from '../../views/Environments/ColoredEnvironmentStatusComponent.js';
+import { h } from '/js/src/index.js';
+import { STATUS_ACRONYMS } from '../../domain/enums/statusAcronyms.js';
+
+/**
+ * Returns a legend for the environment status history acronyms.
+ *
+ * @return {Component} the resulting environment status history legend component
+ */
+export const environmentStatusHistoryLegendComponent = () =>
+    h('', [
+        h('h5', 'Status History Legend'),
+        Object.keys(STATUS_ACRONYMS).map((status) =>
+            h('.flex-row.justify-between', [
+                h('', status),
+                h('', STATUS_ACRONYMS[status]),
+            ])),
+    ]);
 
 /**
  * Returns a tooltip legend for the environment status history acronyms.
  *
- * @param {Component} trigger the element which will display the status history legend when hovered
+ * @param {{status: string, acronym: string}[]} statusHistory list of history items (full status and their acronym)
  * @return {Component} the resulting environment status history component
  */
-export const environmentStatusHistoryComponent = (trigger) =>
-    tooltip(trigger, environmentStatusHistoryLegendComponent());
+export const environmentStatusHistoryComponent = (statusHistory) => tooltip(
+    h(
+        '.flex-row',
+        statusHistory.map((value, index) =>
+            h('.flex-row', [
+                coloredEnvironmentStatusComponent(value.status, value.acronym),
+                h('', index < statusHistory.length - 1 ? '-' : ''),
+            ])),
+    ),
+    environmentStatusHistoryLegendComponent(),
+);
