+++ lib/public/views/Environments/format/displayEnvironmentStatusHistoryOverview.js
@@ -12,7 +12,6 @@
  */
 import { h } from '/js/src/index.js';
 import { STATUS_ACRONYMS } from '../../../domain/enums/statusAcronyms.js';
-import { coloredEnvironmentStatusComponent } from '../ColoredEnvironmentStatusComponent.js';
 import { environmentStatusHistoryComponent } from '../../../components/environments/environmentStatusHistoryComponent.js';
 
 /**
@@ -26,17 +25,10 @@ export const displayEnvironmentStatusHistoryOverview = (environment) => {
 
     const statusHistory = historyItems
         .filter(({ status }) => status in STATUS_ACRONYMS)
-        .map(({ status }) => ({ status, content: STATUS_ACRONYMS[status] }));
+        .map(({ status }) => ({ status, acronym: STATUS_ACRONYMS[status] }));
 
     return h(
         '.flex-row',
-        environmentStatusHistoryComponent(h(
-            '.flex-row',
-            statusHistory.map((value, index) =>
-                h('.flex-row', [
-                    coloredEnvironmentStatusComponent(value.status, value.content),
-                    h('', index < statusHistory.length - 1 ? '-' : ''),
-                ])),
-        )),
+        environmentStatusHistoryComponent(statusHistory),
     );
 };
